rpms: remove some unneccesary calls to get_selected_screen for performance reasons
We just use the variable as is. ratpoison will export it for us. On
the other hand this breaks calling rpms from a terminal, since then we
would need to use "ratpoison -c 'getenv rpms_selected_screen'" to be
sure to have the latest value, and not the value at the time the
terminal was spawned by ratpoison.