Introduce swap screen prev/next commands (hacky)
This is a rather hacky implementation that just uses repeated screen
switches and desktop_other_global to switch the active workspace
between the two screens, by using the last used (global) workspace as
a temporary.

A more proper, and likely faster, implementation would probably
manipulate the various variables and commands directly, using a temp
variables to store the frame layouts of the screens being swapped.