ratpoisonrc: Make the volume control settings configurable
Allows to customize the volume controls according to needs on
different machines.

Change the regex used to extract the volume from the output of amixer
so that it can handle when there is no dB info (like in the case with
pulseaudio).

Also add rp_ prefix to relevant environment variables (consider doing
this in more places).