rpms: Use get_selected_screen function everywhere instead of env var
The env var only gets updated by the switchscreen hook. This causes
races, and since the global lock the switchscreen hooks won't run
until after the current rpms command completes anyway. This caused
troubles with the current (hacky) implementations of
swap_screen_prev/swap_screen_next as todesktop would see the
non-updated value of rpms_selected_screen.

rpms_selected_screen is currently only used to be able to set
rpms_previous_screen and for commands in ratpoisonrc that want a
convenient way to know the current screen (and for which the problem
above doesn't really apply as they are external to rpms).